requirements.txt
# Data Collection 
requests==2.31.0
scrapy==2.12.0
selenium==4.27.1

# Security - using pycryptodome instead of the deprecated pycrypto
pycryptodome>=3.15

# PDF Processing
unstructured==0.11.0
unstructured_pytesseract==0.3.15
unstructured-inference==0.6.6
pdfminer.six==20240706
pdf2image==1.17.0
opencv-python==4.11.0.86

# Document Processing
python-docx>=0.8.11

# Core data science packages
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.6.1

# NLP and text processing
nltk==3.9.1
gensim==4.3.3
langdetect==1.0.9
transformers==4.39.3
datasets==2.18.0

# Deep learning
torch==2.2.1

# Database and vector search
pgvector==0.2.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.28

# Visualization
matplotlib==3.10.1
seaborn==0.12.0

# Utilities
python-dotenv==1.0.1
tqdm==4.67.1
ipykernel==6.29.5
ipywidgets==8.1.5
colorlog>=6.7.0

# Development tools
ipdb==0.13.13